Resolve fully-uppercase builtin identifiers
Some users try to use uppercase function names like LOG as they are used
to Excel functions, which don't work in fend.

Resolve fully-uppercase builtin identifiers if they are not parsed as
units. This ensures that if, for example, "I" becomes a unit in the
future, it will be resolved as the unit - not the imaginary number.