Keep percentage unit during simplification
This is very similar to #165 which fixed #164, which was effectively
reverted in ecfe3b3dcc6efac465c62f8bf57f8e1b789c991a.

This maintains the "alias" logic of the latter, but special cases
percentages like the former to ensure that percentages are explicitly
given with a query like `11/12*100%`.

Explicitly converting to and from percentages is still possible:

```
> 0.5 to %
50%
> 50% to unitless
0.05
```